pyproject.toml
[project]
name = "kis-ai-trader"
version = "0.1.0"
description = "AI-powered multi-agent stock trading system using KIS API"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "pydantic>=2.9",
    "pydantic-settings>=2.6",
    "pykis>=1.0",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30",
    "alembic>=1.14",
    "redis[hiredis]>=5.0",
    "pandas>=2.2",
    "numpy>=1.26",
    "pandas-ta>=0.3",
    "httpx>=0.27",
    "pyyaml>=6.0",
    "structlog>=24.0",
    "python-telegram-bot>=21.0",
    "prometheus-client>=0.21",
    "apscheduler>=3.10",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
    "ruff>=0.7",
    "mypy>=1.12",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
    "ruff>=0.7",
    "mypy>=1.12",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.mypy]
python-version = "3.12"
strict = true
